Insert rules to own styles
Since some add-ons and plugings inject <style>s of their own to the
`document.head`, use-color-change tried to access the last added
<style>'s rules of the `CSSStyleSheet` object.

This might fail w/ a `DOMException`:

```
CSSStyleSheet.cssRules getter: Not allowed to access cross-origin stylesheet
```

This can be observed for example on https://webkit.org/blog, opening the
Web Console and accessing the rules:

```js
document.styleSheets[1].cssRules
```

This commit changes the logic of adding the rules to the target <style>
element. First, check if there's a use-color-change specific <style>
element already in the document's head. If so, use it to insert the CSS
rules. Otherwise, create a new <style> element w/ a use-color-change
identifier, add it to the document's head, and inesrt the CSS rules.

This fixes the `DOMException`.